IOS-4532 Moved ChangeSignType out of its view
