+++ Tangem/UIComponents/TokenPriceChange/TokenPriceChangeView.swift
@@ -50,32 +50,4 @@ extension TokenPriceChangeView {
         case loading
         case loaded(signType: ChangeSignType, text: String)
     }
-
-    enum ChangeSignType: Int, Hashable {
-        case positive
-        case negative
-        case same
-
-        var imageType: ImageType? {
-            switch self {
-            case .positive:
-                return Assets.quotePositive
-            case .negative:
-                return Assets.quoteNegative
-            case .same:
-                return nil
-            }
-        }
-
-        var textColor: Color {
-            switch self {
-            case .positive:
-                return Colors.Text.accent
-            case .negative:
-                return Colors.Text.warning
-            case .same:
-                return Colors.Text.tertiary
-            }
-        }
-    }
 }
